Updates for the Api v3 Phase II changes that have been deployed
Mainly, commands that could return a plain string must now use [0, <string>] to distinguish them from `st` values.  A plain string returned is now always an `st`.

also a couple of updates from review of changes to NoD code (smaller sizes for lengths stored for undecryptable nodes, logging when nodes can't be deserialized)

+++ src/node.cpp
@@ -519,13 +519,13 @@ bool Node::serialize(string* d)
     // Write data necessary to thaw encrypted nodes.
     if (attrstring)
     {
-        // Write node key data.  These can be quite long, and can be in many shares.  Use 64 bit len
-        uint64_t length1 = nodekeydata.size();
+        // Write node key data.  These can be quite long, and can be in many shares.  Use 32 bit len
+        uint32_t length1 = static_cast<uint32_t>(nodekeydata.size());
         d->append((char*)&length1, sizeof(length1));
         d->append(nodekeydata, 0, size_t(length1));
 
-        // Write attribute string data.   Attributes can be long, too
-        uint64_t length2 = attrstring->size();
+        // Write attribute string data.   Attributes can be long, too, so 32 bit again
+        uint32_t length2 = static_cast<uint32_t>(attrstring->size());
         d->append((char*)&length2, sizeof(length2));
         d->append(*attrstring, 0, size_t(length2));
     }
